+++ src/main/java/gov/usgs/cida/wqp/transform/Transformer.java
@@ -10,6 +10,9 @@ import java.io.OutputStream;
 import java.math.BigDecimal;
 import java.sql.Clob;
 import java.sql.SQLException;
+import java.sql.Timestamp;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 import java.util.Map;
 import java.util.Map.Entry;
 
@@ -107,7 +110,7 @@ public abstract class Transformer extends OutputStream implements ITransformer {
 				rtn = ((Clob) object).getSubString(1, size);
 			} catch (SQLException e) {
 				LOG.info(e.getLocalizedMessage());
-			}
+			}		
 		} else {
 			rtn = object.toString();
 		}