+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - resultPhysChem dataProfile.
 - resultPrimary dataProfile.
 - resultBroad dataProfile.
+- added organization summary
 
 ### Changed
 - Station summary service returns only sites that have summary information.